Purge pre-lemonade navigation
