--- frontend/src/layout/lemonade/RedesignOptIn.scss
@@ -1,16 +1,0 @@
-@import '~/vars.scss';
-
-.RedesignOptIn {
-    display: none;
-    user-select: none;
-    .info {
-        font-weight: 600;
-        color: var(--primary-alt);
-        margin-right: 0.5rem;
-        text-align: right;
-    }
-    @media screen and (min-width: $lg) {
-        display: flex;
-        align-items: center;
-    }
-}
