+++ frontend/src/layout/lemonade/lemonadeLogic.ts
@@ -1,9 +1,25 @@
+import dayjs from 'dayjs'
 import { kea } from 'kea'
+import api from '../../lib/api'
 import { FEATURE_FLAGS } from '../../lib/constants'
 import { featureFlagLogic } from '../../lib/logic/featureFlagLogic'
+import { eventUsageLogic } from '../../lib/utils/eventUsageLogic'
+import { systemStatusLogic } from '../../scenes/instance/SystemStatus/systemStatusLogic'
+import { organizationLogic } from '../../scenes/organizationLogic'
+import { preflightLogic } from '../../scenes/PreflightCheck/logic'
+import { teamLogic } from '../../scenes/teamLogic'
+import { userLogic } from '../../scenes/userLogic'
 import { lemonadeLogicType } from './lemonadeLogicType'
 
-export const lemonadeLogic = kea<lemonadeLogicType>({
+type WarningType =
+    | 'welcome'
+    | 'incomplete_setup_on_demo_project'
+    | 'incomplete_setup_on_real_project'
+    | 'demo_project'
+    | 'real_project_with_no_events'
+    | null
+
+export const lemonadeLogic = kea<lemonadeLogicType<WarningType>>({
     path: ['layout', 'lemonade', 'lemonadeLogic'],
     connect: {
         values: [featureFlagLogic, ['featureFlags']],
@@ -25,6 +41,7 @@ export const lemonadeLogic = kea<lemonadeLogicType>({
         hideToolbarModal: true,
         toggleProjectSwitcher: true,
         hideProjectSwitcher: true,
+        setHotkeyNavigationEngaged: (hotkeyNavigationEngaged: boolean) => ({ hotkeyNavigationEngaged }),
     },
     reducers: {
         isSideBarShownRaw: [
@@ -83,6 +100,12 @@ export const lemonadeLogic = kea<lemonadeLogicType>({
                 hideProjectSwitcher: () => false,
             },
         ],
+        hotkeyNavigationEngaged: [
+            false,
+            {
+                setHotkeyNavigationEngaged: (_, { hotkeyNavigationEngaged }) => hotkeyNavigationEngaged,
+            },
+        ],
     },
     selectors: {
         isSideBarForciblyHidden: [() => [() => document.fullscreenElement], (fullscreenElement) => !!fullscreenElement],
@@ -99,5 +122,115 @@ export const lemonadeLogic = kea<lemonadeLogicType>({
                     : null
             },
         ],
+        systemStatus: [
+            () => [
+                systemStatusLogic.selectors.overview,
+                systemStatusLogic.selectors.systemStatusLoading,
+                preflightLogic.selectors.siteUrlMisconfigured,
+            ],
+            (statusMetrics, statusLoading, siteUrlMisconfigured) => {
+                if (statusLoading) {
+                    return true
+                }
+
+                if (siteUrlMisconfigured) {
+                    return false
+                }
+
+                // On cloud non staff users don't have status metrics to review
+                const hasNoStatusMetrics = !statusMetrics || statusMetrics.length === 0
+                if (hasNoStatusMetrics && preflightLogic.values.preflight?.cloud && !userLogic.values.user?.is_staff) {
+                    return true
+                }
+
+                // if you have status metrics these three must have `value: true`
+                const aliveMetrics = ['redis_alive', 'db_alive', 'plugin_sever_alive']
+                const aliveSignals = statusMetrics
+                    .filter((sm) => sm.key && aliveMetrics.includes(sm.key))
+                    .filter((sm) => sm.value).length
+                return aliveSignals >= aliveMetrics.length
+            },
+        ],
+        updateAvailable: [
+            (selectors) => [
+                selectors.latestVersion,
+                selectors.latestVersionLoading,
+                preflightLogic.selectors.preflight,
+            ],
+            (latestVersion, latestVersionLoading, preflight) => {
+                // Always latest version in multitenancy
+                return (
+                    !latestVersionLoading &&
+                    !preflight?.cloud &&
+                    latestVersion &&
+                    latestVersion !== preflight?.posthog_version
+                )
+            },
+        ],
+        demoWarning: [
+            () => [organizationLogic.selectors.currentOrganization, teamLogic.selectors.currentTeam],
+            (organization, currentTeam): WarningType => {
+                if (!organization) {
+                    return null
+                }
+
+                if (
+                    organization.setup.is_active &&
+                    dayjs(organization.created_at) >= dayjs().subtract(1, 'days') &&
+                    currentTeam?.is_demo
+                ) {
+                    return 'welcome'
+                } else if (organization.setup.is_active && currentTeam?.is_demo) {
+                    return 'incomplete_setup_on_demo_project'
+                } else if (organization.setup.is_active) {
+                    return 'incomplete_setup_on_real_project'
+                } else if (currentTeam?.is_demo) {
+                    return 'demo_project'
+                } else if (currentTeam && !currentTeam.ingested_event) {
+                    return 'real_project_with_no_events'
+                }
+                return null
+            },
+        ],
     },
+    loaders: {
+        latestVersion: [
+            null as string | null,
+            {
+                loadLatestVersion: async () => {
+                    const versions = (await api.get('https://update.posthog.com')) as VersionType[]
+                    for (const version of versions) {
+                        if (
+                            version?.release_date &&
+                            dayjs
+                                .utc(version.release_date)
+                                .set('hour', 0)
+                                .set('minute', 0)
+                                .set('second', 0)
+                                .set('millisecond', 0) > dayjs()
+                        ) {
+                            // Release date is in the future
+                            continue
+                        }
+                        return version.version
+                    }
+                    return null
+                },
+            },
+        ],
+    },
+    listeners: ({ actions }) => ({
+        setHotkeyNavigationEngaged: async ({ hotkeyNavigationEngaged }, breakpoint) => {
+            if (hotkeyNavigationEngaged) {
+                eventUsageLogic.actions.reportHotkeyNavigation('global', 'g')
+                await breakpoint(3000)
+                actions.setHotkeyNavigationEngaged(false)
+            }
+        },
+    }),
+    events: ({ actions }) => ({
+        afterMount: () => {
+            actions.loadLatestVersion()
+        },
+    }),
 })
