--- frontend/src/layout/navigation/index.ts
@@ -1,5 +1,0 @@
-import { MainNavigation } from './MainNavigation'
-import { TopNavigation } from './TopNavigation'
-import { DemoWarnings } from './DemoWarnings'
-
-export { MainNavigation, TopNavigation, DemoWarnings }
