+++ frontend/src/scenes/App.tsx
@@ -3,7 +3,6 @@ import { kea, useMountedLogic, useValues } from 'kea'
 import { Layout } from 'antd'
 import { ToastContainer, Slide } from 'react-toastify'
 import { preflightLogic } from './PreflightCheck/logic'
-import { MainNavigation, TopNavigation, DemoWarnings } from '~/layout/navigation'
 import { BillingAlerts } from 'lib/components/BillingAlerts'
 import { userLogic } from 'scenes/userLogic'
 import { sceneLogic } from 'scenes/sceneLogic'
@@ -14,12 +13,13 @@ import { BackTo } from 'lib/components/BackTo'
 import { appLogicType } from './AppType'
 import { models } from '~/models'
 import { FEATURE_FLAGS } from 'lib/constants'
-import { CloudAnnouncement } from '~/layout/navigation/CloudAnnouncement'
 import { teamLogic } from './teamLogic'
 import { LoadedScene } from 'scenes/sceneTypes'
 import { SideBar } from '../layout/lemonade/SideBar/SideBar'
 import { appScenes } from 'scenes/appScenes'
 import { Breadcrumbs } from '../layout/lemonade/Breadcrumbs/Breadcrumbs'
+import { TopBar } from '../layout/lemonade/TopBar'
+import { DemoWarnings } from '../layout/lemonade/DemoWarnings/DemoWarnings'
 
 export const appLogic = kea<appLogicType>({
     path: ['scenes', 'App'],
@@ -112,7 +112,6 @@ function Models(): null {
 function AppScene(): JSX.Element | null {
     const { user } = useValues(userLogic)
     const { activeScene, params, loadedScenes, sceneConfig } = useValues(sceneLogic)
-    const { featureFlags } = useValues(featureFlagLogic)
     const { showingDelayedSpinner } = useValues(appLogic)
 
     const SceneComponent: (...args: any[]) => JSX.Element | null =
@@ -133,7 +132,7 @@ function AppScene(): JSX.Element | null {
     if (sceneConfig?.plain) {
         return (
             <Layout style={{ minHeight: '100vh' }}>
-                {!sceneConfig.hideTopNav && <TopNavigation />}
+                {!sceneConfig.hideTopNav && <TopBar />}
                 <SceneComponent user={user} {...params} />
                 {toastContainer}
             </Layout>
@@ -143,10 +142,7 @@ function AppScene(): JSX.Element | null {
     const layoutContent = activeScene ? (
         <Layout.Content className="main-app-content" data-attr="layout-content">
             {!sceneConfig?.hideDemoWarnings && <DemoWarnings />}
-            {featureFlags[FEATURE_FLAGS.CLOUD_ANNOUNCEMENT] && !(true || featureFlags[FEATURE_FLAGS.LEMONADE]) ? (
-                <CloudAnnouncement message={String(featureFlags[FEATURE_FLAGS.CLOUD_ANNOUNCEMENT])} />
-            ) : null}
-            {(true || featureFlags[FEATURE_FLAGS.LEMONADE]) && <Breadcrumbs />}
+            <Breadcrumbs />
             <BillingAlerts />
             <BackTo />
             <SceneComponent user={user} {...params} />
@@ -155,20 +151,10 @@ function AppScene(): JSX.Element | null {
 
     return (
         <>
-            {true || featureFlags[FEATURE_FLAGS.LEMONADE] ? (
-                <Layout style={{ minHeight: '100vh' }}>
-                    {!sceneConfig?.hideTopNav && <TopNavigation />}
-                    <SideBar>{layoutContent}</SideBar>
-                </Layout>
-            ) : (
-                <Layout>
-                    <MainNavigation />
-                    <Layout style={{ minHeight: '100vh' }}>
-                        {!sceneConfig?.hideTopNav && <TopNavigation />}
-                        {layoutContent}
-                    </Layout>
-                </Layout>
-            )}
+            <Layout style={{ minHeight: '100vh' }}>
+                {!sceneConfig?.hideTopNav && <TopBar />}
+                <SideBar>{layoutContent}</SideBar>
+            </Layout>
             {toastContainer}
             <UpgradeModal />
         </>
