+++ frontend/src/scenes/authentication/InviteSignup.scss
@@ -194,3 +194,25 @@
         }
     }
 }
+
+.whoami {
+    height: 100%;
+    display: flex;
+    align-items: center;
+
+    .details {
+        flex-direction: column;
+        font-weight: bold;
+        align-items: flex-start;
+        padding: 0 8px;
+
+        span {
+            display: block;
+        }
+
+        span:last-of-type {
+            font-size: 12px;
+            color: $text_muted;
+        }
+    }
+}
